feat: split ClusterItem into own file and add delayed hover tooltip
- Extract ClusterItem into cluster-list/cluster-item.tsx (one component per file)
- Add 500ms framer motion delay before tooltip appears on hover
- Add overflow-hidden to tooltip to prevent text overflow during grow animation
- Use React state for button hover highlight instead of CSS :hover to handle portal overlay
